requirements.txt
anthropic>=0.25.0
streamlit>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=8.0.0
langsmith>=0.1.0
